+++ tsup.config.ts
@@ -5,6 +5,7 @@ export default defineConfig({
     index: "src/index.ts",
     core: "src/core.ts",
     frame: "src/frame.ts",
+    "web-extension": "src/web-extension.ts",
   },
   format: ["esm"],
   dts: true,